# src/quads/server/swagger.yaml
openapi: 3.0.0
info:
  version: '2.2.6'
  title: 'QUADS'
  description: |-
    QUADS automates the future scheduling, end-to-end provisioning and delivery of bare-metal servers and networks.

    This is it's REST API implementation.

    Some useful links:
    - [Quads repository](https://github.com/quadsproject/quads)
    - [Quads blog](https://quads.dev)
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: 'https://quads.example.com/api/v3/'
...
paths:
...
  /schedules/:
    get:
      summary: Returns a list with all the defined schedules
      tags:
        - Schedules
      responses:
        '200':
          description: Schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new schedule
      description: >
        When the cloud's active assignment is a self-schedule, start is set to
        now and end is derived from selfservice.yml settings. Otherwise start
        and end are required in the request body.
      tags:
        - Schedules
      requestBody:
        description: Body request with *schedule* details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hostname
                - cloud
              properties:
                hostname:
                  type: string
                  example: host.example.com
                cloud:
                  type: string
                  example: cloud04
                start:
                  type: string
                  example: 2024-06-02 22:00
                end:
                  type: string
                  example: 2024-07-02 22:00
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - host not available or assignment not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: [ ]
  /schedules/{schedule_id}:
    get:
      summary: Returns a schedule by id
      tags:
        - Schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update an existing schedule
      tags:
        - Schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      requestBody:
        description: Body request with *schedule* details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: Schedule
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: [ ]

    delete:
      summary: Delete schedule by id
      tags:
        - Schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted Schedule
      security:
        - BearerAuth: [ ]
  /schedules/current:
    get:
      summary: Returns all current schedules
      tags:
        - Schedules
      parameters:
        - name: date
          in: query
          description: Date for current
          required: false
          schema:
            type: string
        - name: host
          in: query
          description: Hostname to filter by
          required: false
          schema:
            type: string
        - name: cloud
          in: query
          description: Cloud name to filter by
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Schedule
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/future:
    get:
      summary: Returns all future schedules
      tags:
        - Schedules
      parameters:
        - name: host
          in: query
          description: Hostname to filter by
          required: false
          schema:
            type: string
        - name: cloud
          in: query
          description: Cloud name to filter by
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Schedule
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/hosts_range:
    get:
      summary: Returns a dictionary with hosts as keys and schedules as values on a given range of dates
      tags:
        - Schedules
      parameters:
        - name: start
          in: query
          description: start date
          required: false
          schema:
            type: string
        - name: end
          in: query
          description: end date
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ScheduleHostsRange
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleHostsRange'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/stats/build_delta:
    get:
      summary: Returns build time delta statistics for schedules
      tags:
        - Schedules
      responses:
        '200':
          description: Build delta statistics
          content:
            application/json:
              schema:
                type: object
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/stats/utilization:
    get:
      summary: Returns utilization statistics for a date range
      tags:
        - Schedules
      parameters:
        - name: start
          in: query
          description: Start date
          required: true
          schema:
            type: string
        - name: end
          in: query
          description: End date
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Utilization statistics
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad Request - start and end required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    ScheduleHostsRange:
      type: object
      properties:
        host:
          type: string
          example: host.example.com
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
...
    Schedule:
      type: object
      properties:
        start:
          type: string
          description: Accepted as naive input; responses serialize as RFC 1123 GMT (real UTC)
          example: 2022-02-02T22:00
        end:
          type: string
          description: Accepted as naive input; responses serialize as RFC 1123 GMT (real UTC)
          example: 2022-04-02T22:00
        build_start:
          type: string
          description: Accepted as naive input; responses serialize as RFC 1123 GMT (real UTC)
          example: 2022-02-02T22:00
        build_end:
          type: string
          description: Accepted as naive input; responses serialize as RFC 1123 GMT (real UTC)
          example: 2022-04-02T22:00
        assignment_id:
          type: integer
          example: 12
        host_id:
          type: integer
          example: 1
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 400
          format: int32
        message:
          type: string
          example: Something went horribly wrong.
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer